+++ guides/tx-creation-and-signing/multisig.js
@@ -1,100 +1,11 @@
 const { cryptography,codec } = require('@liskhq/lisk-client');
+const { transactionSchema, registerMultisignatureParamsSchema } = require('./schemas');
 const readline = require("readline");
+
+let privateKeyStr;
 
 // Public identifier of the account that is converted to multi-sig
 const pubkey = "e57a23f897b13bdeef27439bb9f4e29ac0828018d27d6b39ade342879928b46a";
-
-// Schemas
-const registerMultisignatureParamsSchema = {
-	$id: '/test/auth/command/regMultisig',
-	type: 'object',
-	properties: {
-		numberOfSignatures: {
-			dataType: 'uint32',
-			fieldNumber: 1,
-			minimum: 1,
-			maximum: 64,
-		},
-		mandatoryKeys: {
-			type: 'array',
-			items: {
-				dataType: 'bytes',
-				minLength: 32,
-				maxLength: 32,
-			},
-			fieldNumber: 2,
-			minItems: 0,
-			maxItems: 64,
-		},
-		optionalKeys: {
-			type: 'array',
-			items: {
-				dataType: 'bytes',
-				minLength: 32,
-				maxLength: 32,
-			},
-			fieldNumber: 3,
-			minItems: 0,
-			maxItems: 64,
-		},
-		signatures: {
-			type: 'array',
-			items: {
-				dataType: 'bytes',
-				minLength: 64,
-				maxLength: 64,
-			},
-			fieldNumber: 4,
-		},
-	},
-	required: ['numberOfSignatures', 'mandatoryKeys', 'optionalKeys', 'signatures'],
-};
-const transactionSchema = {
-	$id: '/lisk/transaction',
-	type: 'object',
-	required: ['module', 'command', 'nonce', 'fee', 'senderPublicKey', 'params'],
-	properties: {
-		module: {
-			dataType: 'string',
-			fieldNumber: 1,
-			minLength: 1,
-			maxLength: 32,
-		},
-		command: {
-			dataType: 'string',
-			fieldNumber: 2,
-			minLength: 1,
-			maxLength: 32,
-		},
-		nonce: {
-			dataType: 'uint64',
-			fieldNumber: 3,
-		},
-		fee: {
-			dataType: 'uint64',
-			fieldNumber: 4,
-		},
-		senderPublicKey: {
-			dataType: 'bytes',
-			fieldNumber: 5,
-			minLength: 32,
-			maxLength: 32,
-		},
-		params: {
-			dataType: 'bytes',
-			fieldNumber: 6,
-		},
-		signatures: {
-			type: 'array',
-			items: {
-				dataType: 'bytes',
-			},
-			fieldNumber: 7,
-		},
-	},
-};
-
-let privateKeyStr;
 
 // Parameters
 let paramsStr = {
@@ -106,6 +17,7 @@ let paramsStr = {
 	],
 	"signatures": []
 };
+
 let params = {
 	...paramsStr,
 	optionalKeys: [
@@ -113,9 +25,6 @@ let params = {
 		Buffer.from("dfbe4e3999138d62047c23f61f222a91b24d9d056db055be24f9ab6d95d7aa78","hex")
 	]
 }
-
-console.log("=============== params ===============");
-console.log(params);
 
 const rl = readline.createInterface({
 	input: process.stdin,
@@ -129,6 +38,7 @@ if (process.argv.length < 3) {
 
 console.log("Please only proceed to sign the registration message if you confirm the correctness of the following parameters for the multi-signature registration:");
 console.log(paramsStr);
+
 rl.question("Confirm parameters with 'yes'", function(confirmed) {
 	confirmed = confirmed.toLowerCase();
 	if (confirmed == "yes" || confirmed == "y") {
@@ -154,13 +64,6 @@ rl.question("Confirm parameters with 'yes'", function(confirmed) {
 		const tag = 'LSK_RMSG_';
 		const privateKey = Buffer.from(privateKeyStr,"hex");
 
-	/*	const decodedBaseTransaction = codec.codec.decode(
-			transactionSchema,
-			unsignedMultiSigTransaction,
-		);
-
-		console.log(decodedBaseTransaction);*/
-
 		const signature = cryptography.ed.signDataWithPrivateKey(
 			tag,
 			chainID,
@@ -168,6 +71,7 @@ rl.question("Confirm parameters with 'yes'", function(confirmed) {
 			privateKey,
 		);
 
+		console.log("Signature:");
 		console.log(signature);
 
 		params.signatures.push(signature);
